ganti int jadi float

+++ src/main/java/ken/backend/algoritma/BillProcess.java
@@ -5,5 +5,5 @@ import ken.backend.kelas.bill.BillItem;
 import java.util.Map;
 
 public interface BillProcess{
-    int process(Map<Integer, BillItem> listBarang);
+    float process(Map<Integer, BillItem> listBarang);
 }
